cleanup const for imports
