+++ server/package.json
@@ -22,11 +22,13 @@
     "pg": "^8.11.3"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/jsonwebtoken": "^9.0.6",
     "@types/morgan": "^1.9.9",
     "@types/node": "^20.11.25",
+    "@types/pg": "^8.11.2",
     "morgan": "^1.10.0",
     "nodemon": "^3.1.0",
     "ts-node": "^10.9.2",
